First stage people/all and people/add is working

+++ domain/src/main/kotlin/app/mobius/domain/entity/Profile.kt
@@ -63,7 +63,7 @@ data class Phone(
         @Id @GeneratedValue @Column(name = "phone_uuid") val phoneUUID: UUID? = null,
         @Column(name = "code_country") var codeCountry: String,
         @Column(name = "code_area") var codeArea: String,
-        @Column(name = "number") var number: Int
+        @Column(name = "number") var number: Long
 ) {
         constructor() : this(codeCountry = "", codeArea = "", number = -1)
 }
